EduRAG/Streamlit cloud/requirements.txt
streamlit
huggingface-hub
sentence-transformers
faiss-cpu
numpy
pymupdf
